Do not remove DC computer account
Some potential issues with automating this step (such as passing in the
ComputerName of an existing, working DC), and removing the account this
way doesn't seem to sufficiently clean up AD so that the DC can be
re-added (which was the goal of removing the account in the first
place). For the time-being, if a DC needs to be rebuilt, records of the
DC in domain will need to be cleaned up manually before proceeding.