Add AmiId Parameter to CloudFormation Templates
Added AmiId parameter to the guac, rdsh, and rdgw compound and the
autoscale_elb.element templates. Also added in the parameter section
of the autoscale nested template in the compound temnplate. Added the
UseAmiId and NotToUseAmiId conditions to the autoscale templates of the
three components to test whether to use the AmiId for image. If blank
the result from the lambda function will be used for the AMI.
A condition statement was added in the lambda resource section so that
if the AmiId paramter is entered the resource is ignored. In the
Launch Config section a test of the UseAmiId was added. If true the
AMI enetered will be used else the result of the lambda function used.