Removes connection broker role from rdsh instances
Also adds some cleanup steps, so any non-responsive rdsh instances
are removed from the deployment.

Installing the connection broker role to the rdsh instances set up
the circumstances for the deployment to reach an irrecoverable
state:

* New servers cannot be added to the deployment if any connection
broker is unavailable
* The connection broker role cannot be removed from unavailable
servers if any _other_ connection brokers are unavailable.

Therefore, with the connection broker role installed to the rdsh
instances, if the autoscaling group terminated two instances then
the deployment would reach an irrecoverable state.

Until another approach is found, or it becomes possible to remove
unavailable connection brokers from the deployment, it seems best
not to autoscale the connection broker role.