Disables the autoscaling azrebalance process explicitly
When the desired capacity is greater than the number of AZs associated
to the auto-scaling group, and when manually managing the active/standby
status of each instance, a potential race condition occurs. Here are the
conditions:

- One or more AZs have multiple instances
- All instances place themselves into standby
- Each instance progresses at its own pace
- Multiple instances in a single AZ exit standby
- At least one AZ contains 0 active instances

In this scenario, an AZ imbalance occurs and the auto-scaling group will
then terminate and launch instances to restore the AZ balance. Of course,
the only reason for the imbalance is because an instance has yet to exit
standby. When that instance becomes active again, the imbalance would
have been restored anyway, making the terminate/launch activity pointless
and actually *increasing* the total time for the pool to stabilize.

The race condition may also occur when the instances are entering standby,
if the AZRebalance process executes whenever zero instances are active
in one AZ and two or more instances are active in another AZ.

The patch is an attempt to address this race condition by disabling the
AZRebalance process entirely.